requirements.txt
fastapi
uvicorn
graphene
pymongo
graphene-pydantic
pydantic
starlette_graphene3
strawberry
magic
python-decouple